move cache blocks command to job
