+++ app/Console/Kernel.php
@@ -7,7 +7,6 @@ namespace App\Console;
 use App\Console\Commands\BuildForgingStats;
 use App\Console\Commands\CacheAddressStatistics;
 use App\Console\Commands\CacheAnnualStatistics;
-use App\Console\Commands\CacheBlocks;
 use App\Console\Commands\CacheCurrenciesData;
 use App\Console\Commands\CacheDelegateAggregates;
 use App\Console\Commands\CacheDelegatePerformance;
@@ -84,8 +83,6 @@ final class Kernel extends ConsoleKernel
         $schedule->command(CacheDelegateProductivity::class)->everyMinute();
 
         $schedule->command(CacheTransactions::class)->everyFiveMinutes();
-
-        $schedule->command(CacheBlocks::class)->everyFiveMinutes();
 
         $schedule->command(CacheAddressStatistics::class)->everyFiveMinutes();
